+++ packages/article-magazine-comment/src/styles/index.android.js
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+import { spacing } from "@times-components/styleguide";
+import nativeStyles from "./native";
+
+const androidStyles = {
+  ...nativeStyles,
+  articleHeadline: {
+    ...nativeStyles.articleHeadline,
+    marginBottom: spacing(3)
+  },
+  metaTablet: {
+    ...nativeStyles.metaTablet,
+    paddingTop: "auto"
+  },
+  standFirst: {
+    ...nativeStyles.standFirst,
+    marginBottom: spacing(2.5)
+  }
+};
+
+const styles = StyleSheet.create({
+  ...androidStyles
+});
+
+export default styles;