+++ packages/image/src/utils/index.js
@@ -1,3 +1,4 @@
+import url from "url";
 import Utils from "./appendToURL";
 
 const appendParamsToQuery = (uriString, paramMap) => {
@@ -5,22 +6,10 @@ const appendParamsToQuery = (uriString, paramMap) => {
     return uriString;
   }
 
-  let finalUri = uriString;
-
-  paramMap.forEach(item => {
-    if (
-      !(
-        (item && finalUri.includes(`?${item.name}`)) ||
-        finalUri.includes(`&${item.name}`)
-      ) &&
-      item.value
-    ) {
-      const separator = finalUri.includes("?") ? "&" : "?";
-      finalUri = `${finalUri}${separator}${item.name}=${item.value}`;
-    }
-  });
-
-  return finalUri;
+  const uri = url.parse(uriString, true);
+  uri.search = undefined;
+  Object.assign(uri.query, paramMap);
+  return url.format(uri);
 };
 
 export default Utils;