+++ packages/provider-queries/src/author.js
@@ -1,18 +1,4 @@
 import { addTypenameToDocument } from "apollo-utilities";
-import gql from "graphql-tag";
+import author from "./author.graphql";
 
-export default addTypenameToDocument(gql`
-  query AuthorQuery($slug: Slug!) {
-    author(slug: $slug) {
-      articles {
-        count
-      }
-      biography
-      hasLeadAssets
-      image
-      jobTitle
-      name
-      twitter
-    }
-  }
-`);
+export default addTypenameToDocument(author);