+++ packages/provider-queries/webpack.config.js
@@ -1,5 +1,25 @@
 /* eslint-disable import/no-extraneous-dependencies */
-module.exports = require("@times-components/webpack-configurator")(
+const path = require("path");
+
+const toBeExported = {
+  entry: {
+    files: path.resolve(__dirname, "./src/article-extras.js")
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(graphql|gql)$/,
+        use: {
+          loader: "graphql-tag/loader"
+        }
+      }
+    ]
+  }
+};
+
+const defaultConfiguration = require("@times-components/webpack-configurator")(
   __dirname,
   "dev"
 );
+
+module.exports = Object.assign(defaultConfiguration, toBeExported);