+++ packages/mock-tpa-server/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.23](https://github.com/newsuk/times-components/compare/@times-components/mock-tpa-server@0.2.22...@times-components/mock-tpa-server@0.2.23) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/mock-tpa-server
+
+
+
+
+
 ## [0.2.22](https://github.com/newsuk/times-components/compare/@times-components/mock-tpa-server@0.2.21...@times-components/mock-tpa-server@0.2.22) (2019-08-14)
 
 **Note:** Version bump only for package @times-components/mock-tpa-server