+++ packages/provider-queries/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.1](https://github.com/newsuk/times-components/compare/@times-components/provider-queries@2.7.0...@times-components/provider-queries@2.7.1) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/provider-queries
+
+
+
+
+
 # [2.7.0](https://github.com/newsuk/times-components/compare/@times-components/provider-queries@2.6.2...@times-components/provider-queries@2.7.0) (2019-08-14)