+++ packages/section/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.17](https://github.com/newsuk/times-components/compare/@times-components/section@1.2.16...@times-components/section@1.2.17) (2019-08-28)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [1.2.16](https://github.com/newsuk/times-components/compare/@times-components/section@1.2.15...@times-components/section@1.2.16) (2019-08-27)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [1.2.15](https://github.com/newsuk/times-components/compare/@times-components/section@1.2.14...@times-components/section@1.2.15) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [1.2.14](https://github.com/newsuk/times-components/compare/@times-components/section@1.2.13...@times-components/section@1.2.14) (2019-08-22)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [1.2.13](https://github.com/newsuk/times-components/compare/@times-components/section@1.2.12...@times-components/section@1.2.13) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
 ## [1.2.12](https://github.com/newsuk/times-components/compare/@times-components/section@1.2.11...@times-components/section@1.2.12) (2019-08-21)
 
 **Note:** Version bump only for package @times-components/section